% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_SMD.R
\name{add_SMD}
\alias{add_SMD}
\title{add_SMD}
\usage{
add_SMD(tbl, location, ref_group, ci, decimals, ci_bracket, ci_sep)
}
\arguments{
\item{tbl}{Must be a gtsummary object of type 'tbl_summary' or 'tbl_svysummary'.}

\item{location}{Can be set to "label" (one SMD is computed for each variable; for categorical variables, a Mahalanobis distance is computed [1]), "level" (an SMD is computed for every level of categorical variables), or "both" (a combination of both "label" and "level"). Default is "label".}

\item{ref_group}{Binary. If TRUE, group 1 is set as a reference group, and SMDs are computed between it and all other groups. If FALSE, every pairwise combination of groups is computed. Default is FALSE.}

\item{ci}{Binary. If TRUE, confidence intervals are added to the SMDs.}

\item{decimals}{Integer. Sepcified the number of decimals to round to for SMDs (and confidence intervals).}

\item{ci_bracket}{Character. Default is set to "{}". Any string can be specified; the first character will be used as the opening bracket for the confidence interval, and the second character as the closing bracket.}

\item{ci_sep}{Character. Default is ", ". Specified the separator characters to use between the lower and the upper confidence bounds.}
}
\value{
Returns a tbl_summary or tbl_svysummary object (same as the input to the "tbl" argument) with added SMDs.
}
\description{
This function takes a tbl_summary or tbl_svysummary object as input and adds columns with standardised mean differences (SMDs) between the "by"-groups in the table. There is the option to compute SMDs between every possible pair of groups or, alternatively, between one reference group and all other groups.
}
\examples{
  \dontrun{library(gtsummary)
  trial \%>\% tbl_summary(by = "trt") \%>\% add_SMD()}
}
\references{
[1]: Yang & Dalton (2012): A unified approach to measuring the effect size between two groups using SAS® (https://support.sas.com/resources/papers/proceedings12/335-2012.pdf)
}
\seealso{
See \code{\link[gtsummary]{tbl_summary}} and \code{\link[gtsummary]{tbl_svysummary}}
}
\author{
Zheer Kejlberg Al-Mashhadi
}
