% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_SMD_v1.R
\name{add_SMD_v1}
\alias{add_SMD_v1}
\title{add_SMD_v1}
\description{
This is an internal Z.gtsummary.addons function
}
\seealso{
See \code{\link{add_SMD}}
}
\author{
Zheer Kejlberg Al-Mashhadi
}
\keyword{internal}
