% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_5_gtsummary.R
\name{round_5_gtsummary}
\alias{round_5_gtsummary}
\title{round_5_gtsummary}
\usage{
round_5_gtsummary(table)
}
\arguments{
\item{table}{Must be a gtsummary object of type 'tbl_summary' or 'tbl_svysummary'.}
}
\value{
Returns a tbl_summary or tbl_svysummary object (same as the input to the "table" argument).
}
\description{
This function takes a tbl_summary or tbl_svysummary object as input and rounds all counts to nearest 5. If counts are between 0-4, the output will be "<5". Similarly, if counts are between n and n-4, the output will be ">{n-5}".
}
\examples{
  \dontrun{library(gtsummary)
  trial \%>\% tbl_summary(by = "trt") \%>\% round_5_gtsummary()}
}
\author{
Zheer Kejlberg Al-Mashhadi
}
