% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_smd_data.R
\name{clean_smd_data}
\alias{clean_smd_data}
\title{clean_smd_data}
\description{
This is an internal Z.gtsummary.addons function
}
\seealso{
See \code{\link[Z.gtsummary.addons::add_SMD]{add_SMD}}
}
\author{
Zheer Kejlberg Al-Mashhadi
}
\keyword{internal}
