\name{core_smd_function}
\alias{core_smd_function}

\title{
core_smd_function
}
\description{
Internal Z.gtsummary.addon function
}

\author{
Zheer Kejlberg Al-Mashhadi
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
See \code{\link[Z.gtsummary.addons::add_SMD]{add_SMD}}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.

