% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_smd_function.R
\name{core_smd_function}
\alias{core_smd_function}
\title{core_smd_function}
\description{
This is an internal Z.gtsummary.addons function
}
\seealso{
See \code{\link[Z.gtsummary.addons::add_SMD]{add_SMD}}
}
\author{
Zheer Kejlberg Al-Mashhadi
}
\keyword{internal}
